requirements.txt
imutils==0.5.4
opencv_contrib_python==4.5.3.56
pigpio==1.78
numpy==1.24.3
gps==3.19
protobuf==3.20.3
pyrealsense2==2.53.1.4623
pyserial==3.5
